SwordAttack
Creación SwordAttack y arreglo de int a double

+++ src/main/java/main/combat/CombatResult.java
@@ -3,16 +3,16 @@ package main.combat;
 import main.state.CharacterState;
 
 public class CombatResult {
-    private final int damage;
+    private final double damage;
     private final CharacterState newState;
     private final boolean isBlocking;
 
-    public CombatResult(int damage, CharacterState newState, boolean isBlocking){
+    public CombatResult(double damage, CharacterState newState, boolean isBlocking){
         this.damage = damage;
         this.newState = newState;
         this.isBlocking = isBlocking;
     }
-    public int getDamage(){
+    public double getDamage(){
         return damage;
     }
     public CharacterState getNewState(){
